analysis script
