+++ scripts/log_analysis/depreciate/deprec-consumer-analysis.py
