+++ scripts/log_analysis/depreciate/deprec-mtnd-consumer-analysis.py
