+++ scripts/multinode-complete-deployment.sh
@@ -446,9 +446,7 @@ for file in $INPUT_GRAPH_FOLDER/*.bs.yaml; do
         {
             cd $DIR_OUTPUT_FINAL || exit 1
             "$SCRIPT_DIR/log_analysis/extractLogs.sh" filebeat* >/dev/null 2>&1
-            python3 "$SCRIPT_DIR/log_analysis/mtnd-consumer-analysis.py" consumer_logs.txt &
-            python3 "$SCRIPT_DIR/log_analysis/mtnd-controller-analysis.py" controller_logs.txt &
-            python3 "$SCRIPT_DIR/log_analysis/mtnd-experience-analysis.py" consumer_logs.txt controller_logs.txt &
+            python3 "$SCRIPT_DIR/log_analysis/mtnd-analysis.py" consumer_logs.txt controller_logs.txt &
             python3 "$SCRIPT_DIR/../experience/generator/graph_visualizor.py" $SCRIPT_DIR/../$file &
             wait
         } &
